requirements.txt
kagglehub>=0.2.5
pandas>=2.2.0
numpy>=1.25.0
sentence-transformers>=2.6.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
tqdm>=4.66.0
pyarrow>=15.0.0
streamlit>=1.32.0
rank-bm25>=0.2.2
matplotlib>=3.8.0

